Prototype for a new lightweight Equation::Connect evaluation
It all comes down to the realization that on connect, we can link
`Var`'s to each other so we don't loose track of these connection.

Remember, `Var` is the only place in icomb that has mutations. And by
linking `Var`'s we loose Interaction Combinator's locality. However,
all is not lost. With careful use of atomic's in Rust, we can make sure
these links are thread-safe.

So Equation::Connect evaluations are fully supported including chained
connections.